fix: Move wrong position shared config
