+++ src/ai/backend/manager/config/scaling_group.py
@@ -1,0 +1,8 @@
+# TODO: Make this more precise type
+from pydantic import BaseModel, ConfigDict
+
+
+class ScalingGroupConfig(BaseModel):
+    model_config = ConfigDict(
+        extra="allow",
+    )
